# docs/api/map.openapi.yaml
openapi: 3.1.0
info:
  title: KApp Map API
  version: 3.6.0
  summary: Schematic campus map and space search for Fundacion Universitaria Konrad Lorenz.
  description: |
    Digital campus map for Fundacion Universitaria Konrad Lorenz (Bogota, Colombia).

    ## What this service is (and is not)

    This map is **not** a geospatial service. There is no basemap, no GPS and no external map
    provider, and a floor is never positioned by coordinates: it is a drawing, described as data
    and drawn by the client, not a set of floor plan images either. What version 3.1 adds is the
    ground a campus stands on - see below.

    ## Version 2: what the campus survey changed

    Version 1 was designed before anybody walked the campus. The survey of the real buildings
    (`docs/map/LEVANTAMIENTO.md`) broke five of its assumptions, and version 2 fixes each:

    | Found on campus | Version 1 | Version 2 |
    | --- | --- | --- |
    | A mezzanine between the first and second floors; an entrance at `P0`; a terrace numbered 5 | Floors addressed by an integer `level` | Floors addressed by a **`code`** (`S1`, `P0`, `P1`, `MEZZ`, `T`); `level` is a decimal that only orders them |
    | Bienestar's wings are west and east, and its doors carry no wing mark | A closed `NORTE`/`SUR`/`CENTRAL` enum, guessed from a `-N`/`-S`/`-C` suffix | **Wings declared per building**, each with the suffix its doors carry, if any |
...

...
    photographs of posted plans) or `VERIFIED` (walked).

...

...

...

...

...

...

...

...

...

...

...

...

...

...

...

...

...

...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /api/map/campuses:

    get:
      x-roles: [ROLE_GUEST, ROLE_STUDENT, ROLE_PROFESSOR, ROLE_STAFF, ROLE_ADMIN]
      operationId: listCampuses
      summary: List campuses
      description: |
        Every campus that has at least one building, with its building count.

        Allowed roles: `ROLE_GUEST`, `ROLE_STUDENT`, `ROLE_PROFESSOR`, `ROLE_STAFF`, `ROLE_ADMIN`.
      tags: [Campuses]
      responses:
        '200':
          description: The campuses, ordered by name.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/CampusSummary' }
              example:
                - { name: Sede Principal, buildingCount: 2 }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/map/campuses/{campus}/ground:

    get:
      x-roles: [ROLE_GUEST, ROLE_STUDENT, ROLE_PROFESSOR, ROLE_STAFF, ROLE_ADMIN]
      operationId: getCampusGround
      summary: The ground around a campus
      description: |
        The city around a campus - its blocks, sidewalks, roadways, medians and named streets, and
        the lots of the blocks the university's buildings stand on - from the city's reference
        map, for drawing under the buildings. Coordinates are
        `[lon, lat]` in WGS 84; lay a building's drawing on them with its `placement`. The campus
        name matches ignoring case and accents. `404` for a campus with no ground yet.

        Allowed roles: `ROLE_GUEST`, `ROLE_STUDENT`, `ROLE_PROFESSOR`, `ROLE_STAFF`, `ROLE_ADMIN`.
      tags: [Campuses]
      parameters:
        - name: campus
          in: path
          required: true
          schema: { type: string, maxLength: 120 }
          example: Sede Principal
        - $ref: '#/components/parameters/IfNoneMatch'
      responses:
        '200':
          description: The ground.
          headers:
            ETag: { $ref: '#/components/headers/ETag' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Ground' }
              example:
                campus: Sede Principal
                source: 'Mapa de Referencia para Bogotá D.C., IDECA - UAECD, IDU y otras entidades. CC BY 4.0.'
                retrieved: '2026-09-27'
                blocks:
                  - [[-74.0612, 4.6485], [-74.0616, 4.6489], [-74.0619, 4.6486], [-74.0612, 4.6485]]
                sidewalks: []
                roadways: []
                medians: []
                streets:
                  - { name: Calle 63, label: CL 63, path: [[-74.0617, 4.6491], [-74.0611, 4.6486]] }
        '304': { $ref: '#/components/responses/NotModified' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/map/campuses/{campus}/structures:

    parameters:
      - name: campus
        in: path
        required: true
        schema: { type: string, maxLength: 120 }
        example: Sede Principal

    get:
      x-roles: [ROLE_GUEST, ROLE_STUDENT, ROLE_PROFESSOR, ROLE_STAFF, ROLE_ADMIN]
      operationId: getCampusStructures
      summary: What else stands on a campus's blocks
      description: |
        Everything on the campus's blocks that is not the university's - a neighbour's building,
        a heritage house with its garden - for drawing around the buildings. An empty list, at
        version `0`, while nothing has been saved. The campus name matches ignoring case and
        accents.

        Allowed roles: `ROLE_GUEST`, `ROLE_STUDENT`, `ROLE_PROFESSOR`, `ROLE_STAFF`, `ROLE_ADMIN`.
      tags: [Campuses]
      responses:
        '200':
          description: The campus's structures.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Structures' }
              example:
                campus: Sede Principal
                version: 1
                updatedAt: '2026-09-28T20:00:00Z'
                structures:
                  - name: Vecino, Calle 62 # 9-46 a 9-60
                    floors: 5
                    basements: 0
                    lot: '008213024001'
                    ring: [[-74.0618, 4.6481], [-74.0617, 4.6483], [-74.0615, 4.6482], [-74.0618, 4.6481]]
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }

    put:
      x-roles: [ROLE_ADMIN]
      operationId: saveCampusStructures
      summary: Replace what else stands on a campus's blocks
      description: |
        The **complete** list, as the portal's block editor saves it: a structure missing from it
        is gone. `version` is the list's version as it was read, `0` while nothing has been saved;
        when somebody saved it since, the save is refused with `409` and nothing changes - reload
        and reapply. Every outline must close, `400` otherwise.

        Allowed roles: `ROLE_ADMIN` only.
      tags: [Campuses]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/StructuresRequest' }
            example:
              version: 1
              structures:
                - name: Casa de Francisco de Paula Vélez
                  floors: 1
                  basements: 0
                  lot: '008213024004'
                  ring: [[-74.0620, 4.6484], [-74.0619, 4.6486], [-74.0617, 4.6485], [-74.0620, 4.6484]]
      responses:
        '200':
          description: The list as saved, with its next version.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Structures' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/map/campuses/{campus}/survey:

    parameters:
      - name: campus
        in: path
        required: true
        schema: { type: string, maxLength: 120 }
        example: Sede Principal

    get:
      x-roles: [ROLE_GUEST, ROLE_STUDENT, ROLE_PROFESSOR, ROLE_STAFF, ROLE_ADMIN]
      operationId: getCampusSurvey
      summary: The distances taken round a campus's blocks
      description: |
        Where each wall stands from the curb and how long it is, as taken on site with the portal's
        survey sheet. None, at version `0`, while nothing has been saved. The campus name matches
        ignoring case and accents.

        Allowed roles: `ROLE_GUEST`, `ROLE_STUDENT`, `ROLE_PROFESSOR`, `ROLE_STAFF`, `ROLE_ADMIN`.
      tags: [Campuses]
      responses:
        '200':
          description: The campus's survey.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Survey' }
              example:
                campus: Sede Principal
                version: 4
                updatedAt: '2026-09-29T15:10:00Z'
                measures:
                  - id: bis-09
                    text: '4,80 + 4,96'
                    metres: 9.76
                    note: Planters in the way, measured round them
                    updatedAt: '2026-09-29T15:02:11Z'
                  - id: extra-1
                    label: Gate on Cra 9A, its width
                    text: '3,10'
                    metres: 3.1
                    updatedAt: '2026-09-29T15:09:40Z'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }

    put:
      x-roles: [ROLE_ADMIN]
      operationId: saveCampusSurvey
      summary: Save the distances taken round a campus's blocks
      description: |
        The **complete** list: a distance missing from it is gone. `version` is the survey's
        version as it was read, `0` while nothing has been saved; when it was saved from somewhere
        else since, the save is refused with `409` and nothing changes - read it again and lay the
        changes over it. Two distances with one id, or a negative one, are `400`. `updatedAt` is
        the server's: a distance keeps the time it was taken until its text, total, note or label
        changes.

        Allowed roles: `ROLE_ADMIN` only.
      tags: [Campuses]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SurveyRequest' }
            example:
              version: 4
              measures:
                - id: bis-09
                  text: '4,80 + 4,96'
                  metres: 9.76
      responses:
        '200':
          description: The survey as saved, with its next version.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Survey' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ApiError:
      type: object
      required: [timestamp, status, error, message, path]
      properties:
        timestamp:
          description: When the service produced this error, in UTC.
          type: string
          format: date-time
        status:
          description: The HTTP status, repeated in the body so a logged response is self-contained.
          type: integer
          example: 400
        error:
          description: The status's standard name - `Bad Request`, `Conflict`. For developers.
          type: string
          example: Bad Request
        message:
          description: One sentence a person can act on. This is the line worth surfacing in a UI.
          type: string
        path:
          description: The path that produced the error.
          type: string
        details:
          description: |
            Field-by-field reasons. `field` names what was wrong - in a layout save, with the
            position of the space: `spaces[3].shape` - and `issue` says what about it.
          type: array
          items:
            type: object
            properties:
              field: { type: string }
              issue: { type: string }
...
    Coordinate:
      type: array
      description: A point as GeoJSON writes it, `[lon, lat]`.
      minItems: 2
      maxItems: 2
      items: { type: number }
      example: [-74.0615, 4.6486]
    Ring:
      type: array
      description: The outline of an area, closed - its first point repeated at the end.
      minItems: 4
      items: { $ref: '#/components/schemas/Coordinate' }
    Street:
      type: object
      required: [name, label, path]
      properties:
        name: { type: string, example: Carrera 9 Bis }
        label: { type: string, description: The name as the street signs abbreviate it., example: KR 9BIS }
        path:
          type: array
          description: The street's axis.
          minItems: 2
          items: { $ref: '#/components/schemas/Coordinate' }
    Ground:
      type: object
      description: The city around a campus, from the city's reference map.
      required: [campus, source, retrieved, blocks, sidewalks, roadways, medians, streets]
      properties:
        campus: { type: string, example: Sede Principal }
        source: { type: string, description: "Where the data comes from, to credit wherever it is shown." }
        retrieved: { type: string, format: date, description: The day it was taken from the source. }
        blocks:
          type: array
          description: The city blocks (manzanas) - where private land meets the sidewalk.
          items: { $ref: '#/components/schemas/Ring' }
        sidewalks:
          type: array
          description: The sidewalks (andenes).
          items: { $ref: '#/components/schemas/Ring' }
        roadways:
          type: array
          description: The roadways (calzadas), where cars go.
          items: { $ref: '#/components/schemas/Ring' }
        medians:
          type: array
          description: The medians (separadores) between roadways.
          items: { $ref: '#/components/schemas/Ring' }
        streets:
          type: array
          items: { $ref: '#/components/schemas/Street' }
        lots:
          type: array
          description: The cadastral lots of the blocks the university's buildings stand on - whose land is whose, for surveying a block.
          items: { $ref: '#/components/schemas/Lot' }
    Lot:
      type: object
      description: A cadastral lot.
      required: [code, ring]
      properties:
        code: { type: string, description: "The cadastre's code: the block's nine digits, then the lot's three.", example: '008213024019' }
        ring: { $ref: '#/components/schemas/Ring' }
    Structure:
      type: object
      description: |
        Something on a campus's blocks that is not the university's - a neighbour's building, a
        heritage house with its garden - drawn around the buildings so they read in their place.
      required: [name, floors, basements, ring]
      properties:
        name: { type: string, minLength: 1, maxLength: 120, description: "What it is called, or what it is.", example: Casa de Francisco de Paula Vélez }
        floors: { type: integer, minimum: 0, maximum: 200, description: "How many floors it rises above the street; 0 for ground, like a garden.", example: 1 }
        basements: { type: integer, minimum: 0, maximum: 20, description: How many it goes below., example: 0 }
        lot: { type: string, maxLength: 30, description: "The cadastral lot it stands on, when known.", example: '008213024004' }
        ring: { $ref: '#/components/schemas/Ring' }
    Structures:
      type: object
      description: Everything on a campus's blocks that is not the university's.
      required: [campus, structures, version]
      properties:
        campus: { type: string, example: Sede Principal }
        structures:
          type: array
          items: { $ref: '#/components/schemas/Structure' }
        version: { type: integer, format: int64, description: Bumped by every save. Send it back with the next one., example: 1 }
        updatedAt: { type: string, format: date-time, description: "When it was last saved; absent while nothing has been." }
    StructuresRequest:
      type: object
      description: The whole list, with the version it was read at.
      required: [version, structures]
      properties:
        version: { type: integer, format: int64, minimum: 0, description: "The version the list was read at; 0 while nothing has been saved.", example: 1 }
        structures:
          type: array
          maxItems: 500
          items: { $ref: '#/components/schemas/Structure' }
    SurveyMeasure:
      type: object
      description: |
        One distance taken on site round a campus's blocks, with a tape or a phone's Measure app:
        how far a wall stands from the curb, or how long it is.
      required: [id]
      properties:
        id: { type: string, pattern: '^[a-z0-9][a-z0-9-]{0,39}$', description: "Which distance of the portal's survey plan it is, or an extra one's own id.", example: bis-09 }
        label: { type: string, maxLength: 200, description: "What an extra distance is of; absent for one the plan names.", example: 'Gate on Cra 9A, its width' }
        text: { type: string, maxLength: 120, description: "What was typed, as it was typed: pieces joined with +.", example: '4,80 + 4,96' }
        metres: { type: number, format: double, minimum: 0, maximum: 1000, description: "The total the portal read from the text; absent while it reads none.", example: 9.76 }
        note: { type: string, maxLength: 500, description: "What was seen on site that the sketch does not show.", example: 'Planters in the way, measured round them' }
        recheck: { type: boolean, description: "True while it is asked to be taken again: what was typed stays, for reference, until it is. Absent otherwise.", example: true }
        updatedAt: { type: string, format: date-time, readOnly: true, description: "When it was last changed. Set by the server; ignored in a request." }
    Survey:
      type: object
      description: The distances taken on site round a campus's blocks.
      required: [campus, measures, version]
      properties:
        campus: { type: string, example: Sede Principal }
        measures:
          type: array
          items: { $ref: '#/components/schemas/SurveyMeasure' }
        version: { type: integer, format: int64, description: Bumped by every save. Send it back with the next one., example: 4 }
        updatedAt: { type: string, format: date-time, description: "When it was last saved; absent while nothing has been." }
    SurveyRequest:
      type: object
      description: Every distance taken, the whole list, with the version it was read at.
      required: [version, measures]
      properties:
        version: { type: integer, format: int64, minimum: 0, description: "The version the survey was read at; 0 while nothing has been saved.", example: 4 }
        measures:
          type: array
          maxItems: 500
          items: { $ref: '#/components/schemas/SurveyMeasure' }
...
    CampusSummary:
      type: object
      description: A campus (Sede) and how many buildings the map holds for it.
      required: [name, buildingCount]
      properties:
        name: { type: string, minLength: 1, maxLength: 120, example: Sede Principal }
        buildingCount: { type: integer, minimum: 0, example: 2 }
...
  headers:
    ETag:
      description: The version of this answer. Send it back in `If-None-Match` to be told when it changes. Since 3.6.
      schema: { type: string }
      example: '"3f2a9c14"'
  parameters:
    IfNoneMatch:
      name: If-None-Match
      in: header
      required: false
      description: The `ETag` of the answer the client already has. When nothing changed since, the answer is `304` with no body. Since 3.6.
      schema: { type: string }
      example: '"3f2a9c14"'
...
  responses:
    NotModified:
      description: Nothing changed since the `ETag` sent in `If-None-Match`. No body; keep using what you have.
    BadRequest:
      description: Validation failed
      content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } }
    Unauthorized:
      description: Missing or invalid token
      content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } }
    Forbidden:
      description: Authenticated but not allowed
      content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } }
    NotFound:
      description: Resource not found
      content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } }
    Conflict:
      description: Conflicts with current state
      content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } }